refactor geoResourceByIdProvider
